JoinKeygenViewModel refactored
